Refactor dashboard screen
